--- a/core/variant/dictionary.cpp
+++ b/core/variant/dictionary.cpp
@@ -81,15 +81,7 @@ Variant Dictionary::get_value_at_index(int p_index) const {
 
 Variant &Dictionary::operator[](const Variant &p_key) {
 	if (unlikely(_p->read_only)) {
-		if (p_key.get_type() == Variant::STRING_NAME) {
-			const StringName *sn = VariantInternal::get_string_name(&p_key);
-			const String &key = sn->operator String();
-			if (likely(_p->variant_map.has(key))) {
-				*_p->read_only = _p->variant_map[key];
-			} else {
-				*_p->read_only = Variant();
-			}
-		} else if (likely(_p->variant_map.has(p_key))) {
+		if (likely(_p->variant_map.has(p_key))) {
 			*_p->read_only = _p->variant_map[p_key];
 		} else {
 			*_p->read_only = Variant();
@@ -97,12 +89,7 @@ Variant &Dictionary::operator[](const Variant &p_key) {
 
 		return *_p->read_only;
 	} else {
-		if (p_key.get_type() == Variant::STRING_NAME) {
-			const StringName *sn = VariantInternal::get_string_name(&p_key);
-			return _p->variant_map[sn->operator String()];
-		} else {
-			return _p->variant_map[p_key];
-		}
+		return _p->variant_map[p_key];
 	}
 }