follow up for https://github.com/godotengine/godot/pull/68747
removes conversion from StringName to String in Dictionary

it seems like a simple change, but many places still compare a Variants type with STRING
simply removing the conversion gave me alot of errors, so i did the sane thing and looked at every instance of an equality comparison to `Variant::STRING`, and updated most of them with an `is_string()` method i added to variant:
-- if it accepted String and StringName, i updated it to use `is_string()` (might as well, it's much easier to read)
-- if it was a closed system, (as in no external input, the type is always known to be string), i let it be
-- if it accepted input, (especially from GDScript), i updated it
-- if i couldn't tell for sure if it was a closed system (like in `editor_debugger_inspector.cpp`):
---- if it gave an error for StringName input, i left it (could be caught later)
---- if it silently failed, i updated it

note: Variant does not convert StringName to Color or to NodePath, i'm not sure whether it makes sense to accept StringName, so i let it be for now

fixes https://github.com/godotengine/godot/issues/68834

here are some numbers i got from this example, with different length keys (on a production build):
[TestDictIndexing.zip](https://github.com/godotengine/godot/files/10241791/TestDictIndexing.zip):
```
"%s":
  2.318s: String keys create
 81.679s: StringName keys create
  7.384s: String assign create
  1.973s: String dict String index read
  2.145s: String dict StringName index read
  7.278s: StringName assign create
  2.324s: StringName dict String index read
  1.788s: StringName dict StringName index read

"longer-test-key-%s":
  3.883s: String keys create
 83.304s: StringName keys create
  8.323s: String assign create
  2.406s: String dict String index read
  2.462s: String dict StringName index read
  7.543s: StringName assign create
  2.792s: StringName dict String index read
  1.732s: StringName dict StringName index read
```
with a long enough test key, StringName is faster even when finding a String key:
```
"how-long-could-it-take-to-hash-this-string?-key-%s":
  4.566s: String keys create
 82.376s: StringName keys create
 10.346s: String assign create
  2.949s: String dict String index read
  2.697s: String dict StringName index read
  8.529s: StringName assign create
  3.408s: StringName dict String index read
  1.727s: StringName dict StringName index read
```