@KoBeWi 2022-12-16
I tested this time and it fixes #68834 for real, without affecting usability.