@rune-scape 2024-05-27
it may be because StringName == String doesnt do an in-place comparison like it could, copying the internal String of the StringName (or allocating a new String if not from GDScript) before comparing it leading to an extra ref then deref, ive already made a fix including this change on my branch, ill find some time to test production builds with the settings u mentioned on all 3 variants [before PR, after PR, after PR with performance fix]