@rune-scape 2024-05-30
i changed the test script to autorun and quit and spend less time making keys and more time on the actual test: 
[TestDictIndexing2.zip](https://github.com/godotengine/godot/files/15493077/TestDictIndexing2.zip)

table of percent time of execution compared to before this PR:
![image](https://github.com/godotengine/godot/assets/55224592/d46d0344-8482-4022-8af4-bfd27982471d)

`String keys create`/`StringName keys create` of the large dict test are maybe good anchors for how inconsistent the results are: ~ +/-2%
theres clearly a big slowdown after this PR in `StringName_dict[String]` thats solved by in-place comparisons


theres also the 101.4% `StringName_dict[String]` that 'should' be the exact same or slightly less but thats within the +/-2%

command `scons target=template_release optimise=speed lto=full scu_build=yes`
branch: https://github.com/rune-scape/godot/tree/stringname-dict2

<details><summary>outputs</summary>

- before:
<pre><code>Godot Engine v4.3.beta.custom_build.4bb8c06f6 (2024-05-15 20:42:02 UTC) - https://godotengine.org

key_count/dict_size: 1000000
iterations: 20
  0.47406s: String keys create
  0.41555s: StringName keys create
 12.63348s: String_dict create
 13.42629s: StringName_dict create
  6.92510s: String_dict[String]
  7.23151s: String_dict[StringName]
  7.30472s: StringName_dict[StringName]
  6.96156s: StringName_dict[String]
key_count/dict_size: 50
iterations: 1000000
  0.00003s: String keys create
  0.00002s: StringName keys create
  9.35630s: String_dict create
 11.17116s: StringName_dict create
  4.94667s: String_dict[String]
  6.02697s: String_dict[StringName]
  6.15707s: StringName_dict[StringName]
  4.91070s: StringName_dict[String]</code></pre>
- after:
<pre><code>Godot Engine v4.3.beta.custom_build.fc60dbe4f (2024-05-16 00:01:37 UTC) - https://godotengine.org

key_count/dict_size: 1000000
iterations: 20
  0.47541s: String keys create
  0.39441s: StringName keys create
 12.59615s: String_dict create
 12.29998s: StringName_dict create
  6.98052s: String_dict[String]
  7.20852s: String_dict[StringName]
  6.70885s: StringName_dict[StringName]
  7.46362s: StringName_dict[String]
key_count/dict_size: 50
iterations: 1000000
  0.00003s: String keys create
  0.00002s: StringName keys create
  9.36504s: String_dict create
  9.08053s: StringName_dict create
  4.97574s: String_dict[String]
  6.04046s: String_dict[StringName]
  4.82189s: StringName_dict[StringName]
  6.00519s: StringName_dict[String]</code></pre>
- after w/ in-place comparison:
<pre><code>Godot Engine v4.3.beta.custom_build.5a312e76e (2024-05-28 11:15:25 UTC) - https://godotengine.org

key_count/dict_size: 1000000
iterations: 20
  0.47619s: String keys create
  0.40128s: StringName keys create
 12.61830s: String_dict create
 12.50513s: StringName_dict create
  6.93993s: String_dict[String]
  6.80969s: String_dict[StringName]
  6.80732s: StringName_dict[StringName]
  6.99365s: StringName_dict[String]
key_count/dict_size: 50
iterations: 1000000
  0.00003s: String keys create
  0.00002s: StringName keys create
  9.38292s: String_dict create
  9.08908s: StringName_dict create
  4.93871s: String_dict[String]
  5.04735s: String_dict[StringName]
  4.74848s: StringName_dict[StringName]
  4.97938s: StringName_dict[String]</code></pre>
</details>
<details><summary>perf stats (idk if this is even useful)</summary>

- before:
<pre><code># started on Wed May 29 09:19:01 2024


 Performance counter stats for './bin/godot.linuxbsd.template_release.x86_64.stringname-dict2-0-g4bb8c06f6a --headless ../../Downloads/TestDictIndexing/project.godot':

          99111.20 msec task-clock                       #    1.000 CPUs utilized             
             13853      context-switches                 #  139.772 /sec                      
                72      cpu-migrations                   #    0.726 /sec                      
             75056      page-faults                      #  757.291 /sec                      
      304866650326      cycles                           #    3.076 GHz                       
      577664589537      instructions                     #    1.89  insn per cycle            
      101930137819      branches                         #    1.028 G/sec                     
         179113228      branch-misses                    #    0.18% of all branches           

      99.068539644 seconds time elapsed

      98.726199000 seconds user
       0.313073000 seconds sys
</code></pre>
- after:
<pre><code># started on Wed May 29 09:25:25 2024


 Performance counter stats for './bin/godot.linuxbsd.template_release.x86_64.stringname-dict2-0-gfc60dbe4f0 --headless ../../Downloads/TestDictIndexing/project.godot':

          95753.77 msec task-clock                       #    0.999 CPUs utilized             
             13376      context-switches                 #  139.692 /sec                      
                23      cpu-migrations                   #    0.240 /sec                      
             74266      page-faults                      #  775.593 /sec                      
      294240378895      cycles                           #    3.073 GHz                       
      554352902206      instructions                     #    1.88  insn per cycle            
       96495768265      branches                         #    1.008 G/sec                     
         180022020      branch-misses                    #    0.19% of all branches           

      95.861765139 seconds time elapsed

      95.313911000 seconds user
       0.373233000 seconds sys
</code></pre>
- after w/ in-place comparison:
<pre><code># started on Wed May 29 09:28:41 2024


 Performance counter stats for './bin/godot.linuxbsd.template_release.x86_64.stringname-dict2-0-g5a312e76e4 --headless ../../Downloads/TestDictIndexing/project.godot':

          93036.65 msec task-clock                       #    0.998 CPUs utilized             
             13207      context-switches                 #  141.955 /sec                      
                82      cpu-migrations                   #    0.881 /sec                      
             75291      page-faults                      #  809.262 /sec                      
      286030673926      cycles                           #    3.074 GHz                       
      544972454573      instructions                     #    1.91  insn per cycle            
       94114936119      branches                         #    1.012 G/sec                     
         171548871      branch-misses                    #    0.18% of all branches           

      93.184708944 seconds time elapsed

      92.546298000 seconds user
       0.422159000 seconds sys
</code></pre>
</p>
</details>