@KoBeWi 2024-08-28
Identifier access (i.e. Lua-style) will always use StringNames. Making it use String would force a coversion.