@dalexeev 2024-08-28
You are probably right from a performance perspective. I meant compatibility and consistency issues. In 4.3 there is no difference between Lua- and Python-style dictionaries:

```gdscript
var dict1 := { x = 1 }
var dict2 := { "x": 1 }
print(dict1) # { "x": 1 }
print(dict2) # { "x": 1 }
```