@KoBeWi 2024-08-28
The performance impact is what #68834 is about.